Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 175ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240822/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -3,9 +3,9 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        long m = sc.nextInt();
-        long a = sc.nextInt();
-        long b = sc.nextInt();
+        long m = sc.nextLong();
+        long a = sc.nextLong();
+        long b = sc.nextLong();
 
         long minCnt = m;
         long maxCnt = -1;
